requirements.txt
astropy
comet>=1.1.0
click
voevent-parse>=0.9.0
requests
voeventdb.server
#amiobs>=0.3.2
ipython
Jinja2
python-dateutil
pytz
pytest
pyephem
#imbox
sendgrid
nose
